Improved scaling of Strokes
The thickness of strokes used to be scaled by
a factor of (fx + fy)/2 for scaling factors fx, fy > 0.
This has the disadvantage that for fx != fy undoing the
scaling operation does not restore the correct thickness,
since (fx + fy) != 1/(1/fx + 1/fy) in general. Using the
geometric mean of fx and fy instead of the arithmetic mean
seems to be a better choice.

Signed-off-by: Wilson Brenna <wbrenna@uwaterloo.ca>