Fixed #1139, #1132 by changing the widget which triggers the resize. Fixed a untracked bug, that layoutPages() got the old allocation from the previous run, instead of the desired size allocation.
Refactored most of the files i touched to fix those issues.
Both were related to each other, so i could not split them into two commits.

Added "const" to the Range class.
Added a "noop" to the defines in XournalType, so ClangTidy wont show a warning because of an empty statement

Signed-off-by: Fabian Keßler <fabian_kessler@gmx.de>