Left and right key for page navigation in presentation mode
As the presentation remotes use the left and right keys for page/slide
navigation, map them to previous and next page in presentation
mode. It still preserves the Shift+Left/Right combination in dual page
view. It only masks out the relative scrolling which is not needed in
presentation mode, so thats okay.