correctly handle extensions when exporting
Since #1117, we allow for extensions like ".pdf.xopp". This change led to
a bug where overwriting an existing ".pdf" file in pdf export created a
new file with the extension ".pdf.pdf" instead of overwriting the old
one. This is fixed and additionally all current and future export
filetypes show the same behaviour now.