Prepend iconsDark path after setting default path.
When the "dark theme" option is set, the iconsDark directory should be
the first dir where the icon theme should search for the icon. By
prepending this path after the default icons path has been set, we make
sure that this dir is indeed the first in the list of dirs to look up
icons.

Fixes issue #1746.