Fix (paritally) zooming with scroll wheel
This partially fixes the bug that when zooming with the control key and
mouse wheel, the document scrolls as you zoom.

The variables 'zoomWidgetPosX' and  'zoomWidgetPosX' are being used as
distances rather than positions, so the code does not make sense because
'centerX' and 'centerY' are positions.

The scrolling is still not as smooth as it should be, but it is significantly
better than before.