Minimal fix of ctrl+scroll zoom bug
The 'zoomWidgetPosX' and 'zoomWidgetPosY' are actually offsets, not
positions, and the current code was not taking that into account when
setting the variables.

In addition, I don't believe the size of the toolbars (essentially
'wx' and 'wy' from what I can tell) are relevant. I tested and found
less scrolling when these were removed.