Fix issue #1867.
When moving to the last page of the following PDF document, https://dx.doi.org/10.1109/TSE.1985.232210,
xournal++ hangs (segfaults). You can try to move to the last page either by scrolling
or by setting the page number in the spinpage box, the result will be the same.

After some debugging, I narrowed it down to the latest statement of Layout::updateVisibility():
this->view->getControl()->firePageSelected(mostPageNr). For some reason, this was triggering an
infinite storm of events.

Currently, the updateVisility() is always creating a page selected event, even if there is no
change in page visibility. My solution was to fire the event only if a change occurred
in the visibility.

I used std::optional for the variable mostPageNr. If it was not changed by updateVisibility()'s
algorithm, we do not fire the event that depends upon on mostPageNr's value.