Add option to ignore the first events of the stylus when touching the screen. (#1854)
* Add option to ignore the first events of the stylus when touching the screen to avoid artifacts at the beginnings of strokes. The number of ignored events can be set in the preferences.
Explanation: There are setups with this issue: When the stylus is pressed on the screen, the accuracy and update rate of its recognized position increases. Because of the lower precision before, when hitting the screen the cursor might jump from its previous position to the more precise point (this is especially far if the pen tip was in rapid movement right before touching), while beeing recognized as writing. This can result in artifacts that e.g. make handwritten text very hard to read. To work around this, the first events since the triggering event can be ignored.

* Add requested variable

* Fix handling of deviceClassPressed and implement requested change