Drop fs::exists check for recent files
This will turn recent files into a dumb list that only remembers the
chosen paths, making it more consistent with the behaviour of other
applications. Files that are removed will throw an error when opened,
reminding the user and removing them from the recents list. Files
that are not available when the application starts can be available
later [1].

[1] https://github.com/xournalpp/xournalpp/pull/2304#issuecomment-719408128