MacOS: Update the build environment to gtk 3.24.3 - switch to macports
The MacOS build does not recognize the Wacom pen due to a problem
which is fixed in gtk 3.24.23. I did not manage to create the
build environment with jhbuild so I switched to macports. The Readme
explains how to create the build environment. The gtk-mac-integration
and the gtk-mac-bundler is still used, i.e. this build produces an
application bundle which runs standalone.

See: https://github.com/xournalpp/xournalpp/issues/1772