add new presentation mode for PDF export
This commit adds a new PDF export mode, called "presentation", similar
to what we get from a LaTeX-Beamer document (or Powerpoint).  When
exporting to this mode, each xournalpp page will be rendered to as
many pdf pages as there are layers on the xournalpp page. First, Layer
1 is rendered, then Layer 1 + Layer 2, etc. until all layers are
rendered.

It is accessible by selecting "Save as" in the File menu, and choosing
a PDF export. Then a specific box has to be checked in the Export
Dialog.

Exporting to presentation mode can also be done with the command-line
flag "--export-presentation". For instance:

xournalpp -p output.pdf --export-presentation file.xopp