avoid SVG 1.2 compositing operators
xournalpp paints a white background behind the PDF even for export (but
nor for printing). It does so by painting it after the PDF is rendered,
specifying CAIRO_OPERATOR_DEST_OVER as the compositing operator to get
it below the PDF rendering. When rendering to a SVG 1.2 surface cairo
turns this into a comp-op:dst-over style.

So what's the problem? SVG 1.2 never got accepted as a standard, and
thus it is supported only partially. The resulting SVG does not even
work in inkscape.

Instead, render the white background first and avoid the compositing
operator. This dies not change the resulting image for png outputs but
helps to produce simpler svg.