Layer preview: show PDF background in background layer only
A PDF background lives in the background layer just like any other
background. In the sidebar layer preview, however, PDF backgrounds are
drawn for every layer; all other backgrounds (ruling, image) are drawn
for the background layer only.

Move the corresponding codeblock to drawPage() where all the logic is
about what to draw in the preview, and make it draw PDF backgrounds on
the background layer (and combined page view) only.