sidebar preview: enable toolbar for page preview
If you have a document with bookmarks and switch to the bookmark (index)
preview in the sidebar then the toolbar (containing page up/down/...
buttons) gets disabled, since it does not apply to bookmarks.

Switching from this preview to the layer preview reenables the toolbar,
but switching to the page preview does not.

Make the page preview reenable the toolbar, too.