page preview sidebar: solve toolbar issues
Currently, when you open a file with bookmarks, the index sidebar is
active but the page buttons are visible. They go away only when you
switch to another sidebar and then back to the index sidebar.

If you switch the page (while the index bar is active without toolbar)
the toolbar reappears.

The reason is that `SidebarPreviewPages::pageSelected()` gets called
even when the page preview sidebar is not active.

Teach this method to just record the new selection but not to meddle
with the sidebar view, unless the page preview sidebar is enabled.

This solves both problems (init and page switch).