Sidebar: provide stacked view
Depending on the type of background and the way layers are used,
previews of layers may make more sense either individually or as a
stack.

Provide an alternative sidebar preview "layer stack" which, for each
layer, shows the layer and all layers below, including the background.

Having this as a separate sidebar "page" (tab) make it easy to have
one's favourite view there as well as to switch back and forth quickly.

Technically, this sidebar view is so similar to the existing layer view
that it is implemented as an option for the layer view rather than a new
class.