Fix Selection Tool in New Input System
Previously only `event.button == 1` was handled in the new input system,
but in the old system all non-zero values where considered.
This lead to drawing events not being handled if a button was pressed
in the new input system.
This is fixed in this commit by handling any `event.button > 0`.