PreviewJob: refactor type and background treatment
PreviewJob::drawPage() needs to know the preview type (page/layer) and,
depending on that, the layer. So far, PreviewJob::run() determines that
and passes it using a magic layer number -100 to denote page preview.

Refactor things so that PreviewJob::drawPage() determines this info by
itself. This isolates the actual drawing steps into one method.

Also, this allows to refactor the logic within PreviewJob::drawPage()
clearly: Separate out the Pdf background treatment which is there only
because DocumentView does not know how to do it, and treat the preview
types in a switch/case which allows better separation and extension.

Thus, once DocumentView's methods learn how to deal with Pdf backgrounds
PreviewJob::drawPage() can be simplified easily without hunting for
"hidden" calls to PdfView::drawPage().