Fix isometric grid
For grid painting, the paper is divided into an integer number of cols
and rows. The isometric grid uses only every other point in the grid to
produce the hexagonal lattice. "Isometric dotted" simply stops painting
outside of the paper, but "isometric graph" has to compute the starting
and ending points of the diagonals (and the verticals).

Currently, this works correctly only for the case of an even number of
cols and rows. In all other cases, the right and/or bottom end of the
diagonals are off by 1 grid unit which gives them a wrong slope.
Consequently, not all graph lines intersect as they should.

The simple fix would ensure that cols and rows are always even,
leading to a change in grid layout and a difference between "dotted" and
"graph" versions.

Fix this by taking into account the difference between cols and 2*hdiags
(with hdiags=floor(cols/2) and anologously for the vertical spacing.
This ensures that "dotted" and "graph" versions match exactly.

The existing implementation has the first top left diagonal go from
(0,1) to (1,0) in grid coordinates. As a consequence, there is one more
diagonal in the odd-odd case then in the other cases. We keep in line
with the existing implementation and take care of that extra diagonal
when needed.