Isometric grid: remove default from ini
From its inception, the ini file specified a default grid parameter (20)
which was different from the default in the source code (which coincides
with the default for graph paper, 14.17).

Remove the default from the ini file so that the ini file is free from
any extra parameters.

This does not change how existing files are displayed since they store
any extra parameters for the page templates. It does change how new
files with isometric paper background look: they use the narrower
spacing 14.17 and do not store it in the file, unless a user specifies
the parameter in the ini file explicitely.