position highlight plugin
Until we have a central keybinding factory, this plugin provides a
binding for the position hightlighter. The shortcut Alt+x is chosen so
that on most keyboard layouts it can be done with one hand (the left
one) so that you can keep the stylus or mouse in the other one. This can
serve as a "poor man's laser pointer" quite easily.