allow position highlight without cursor
When the "none" cursor type was introduced in 540ef0df ("Support "no
icon" stylus cursor type #2093", 2020-07-11) it turned off the position
highlighter (as well as the settings) for this mode because of the
intended use case: writing on a "tablet screen" where the physical
stylus indicates the cursor position (without distraction)

There is no technical reason, though, to couple the position highlight
to the cursor type. In addition, the position highlight can serve as a
hinting tool ("laser pointer") when sharing the xournalpp screen, at
least until we have a dedicated tool for that.

So, decouple the settings, which still allows the full distraction free
mode but also pure highlight mode.