fix settings comments placement
The comments for `scrollbarHideType` and for new pages are misplaced in
`settings.xml`, for two different reasons: comments arise as nodes
attached to the node which they comment on in the tree but are output
before the node in a textual representation. Therefore, the code
generating the tree can generate a node comment only after the node it
comments on (1st reason). The usage of macros hides this fact, so when
you circumvent the macros you have to assign the node variable manually,
or else the comment will be attached to the wrong node (2nd reason).
Fix both.