rename settings xml macros
The settings xml tree is generated using `WRITE_*` macros. In fact, they
do not write a settings item out but save/store a node in the xml tree,
before finally writing out a textual representation of the tree. In
particular, `WRITE_COMMENT` adds/attaches a node to the previously saved
node and typically ends up before that in the textual representation.

Rename them to `ATTACH_COMMENT` resp. `SAVE_*` to avoid confusion and to
name them more inline with the underlying functions.