Implement merge-down functionality
The functionality is invoked via
  `LayerController::mergeCurrentLayerDown()`.
Opted not to follow the pattern of the move up/down actions and have
a `...::mergeDown(bool up)` implementation, because in my experience
with other graphical programs (e.g. Krita) the common usage of
merging layers is to merge the upper layer "into" the one below,
which is perfectly sufficient for users.

The core part of the implementation is located in the merge layer
undo/redo action's `redo()` member to avoid doubling the
implementation there.
It is quite simply the for-loop which adds all elements from the
layer above to the layer below and the removal of the layer.

NOTE: There is a caveat to the gui update invoked from the undo/redo
action, as pointed out in the comment in `trigger_ui_update()`:
  | When this implementation is called by the `UndoRedoHandler` the
  | document is locked. Calling `layerChanged` adds a render job which
  | can only be processed when the document is unlocked again, but might
  | have already claimed `Scheduler::jobRunningMutex`.
  | `fireRebuildLayerMenu` will wait for `jobRunningMutex` to be free,
  | so calling `fireRebuildLayerMenu` AFTER `layerChanged` will likely
  | result in a DEADLOCK.

(Implemented with no way to invoke this from the UI)