Create a small context menu for the layer sidebars
This roughly based on the pages sidebar context menu implemented in
xournalpp/xournalpp#1256.

Implements a context menu 'sidebarPreviewLayersContextMenu' separate
from the extisting 'sidebarPreviewContextMenu' (which is for pages
only).
Adds the merging functionality (which is of course the point of this
PR), but also adds the following additional layer operations:
 - move up
 - move down
 - duplicate (copy)
 - delete

The insertion actions (before/after) are currently only defined for
pages and are thus omitted from the context menu.

This adds an abstract class `SidebarBaseContextMenu` which is
supposed to be a base for sidebar context menus. It holds the
pointer to the `GtkWidget`, defines the relevant data structures and
defines an `open()` action which can be set as a callback for a
click. It also requires a member function `setActionsSensitivity()`
which takes a `SidebarActions` set of bits and which derived classes
are to implement to sensitize only the relevant actions' context
menu items whenever this is called.

This then implements this context menu for the layers sidebars in a
derived class called `SidebarLayersContextMenu`. Since we have two
layer sidebars (regular and stacked view) which are objects of the
same class, this class now holds the context menu as a
`std::shared_ptr`. The context menu itself is instantiated before
the instantiation of the sidebars and passed to them as a shared
pointer.