Create a small context menu for the layer sidebars
This roughly based on the pages sidebar context menu implemented in
xournalpp/xournalpp#1256.

I tried to keep the changes minimal and keep it to additions as much
as possible. I implemented a context menu
'sidebarPreviewLayersContextMenu' separate from the extisting
'sidebarPreviewContextMenu' (which is for pages only).
I added the merging functionality (which is of course the point of
this PR) but also added the duplication simply as a quick demo/test
on the button naming for this separate menu.

There is much room for improvement in the consistency between this
and the pages sidebar context menu (beginning with the naming of the
latter) but like I said, I tried to keep the changes minimal.
There also is potential for generifying some of these actions to
the `SidebarPreviewBase` class (maybe then make the string naming
the context menu a param for that?) but I feel like some of the
complexity in dealing with these menus comes from unnecessary
generification between the sidebars, so it should be done with
consideration.

Lastly, since we have to layer sidebars, two `SidebarPreviewLayers`
objects are constructed, and thus I've had to make do with an ugly
workaround using a static int in global scope to keep track of
constructions and destructions and make sure callback registering
and cleanup is done, but not done twice.
A similar issue would exist if the context menu were shared between
pages and layers sidebars, since then we'd have three callback
registrations (if we're not careful) which would result in one click
doing an action three times.

Currently these items are not greyed out when ineligible (e.g.
merge down for lowest layer or background).