Remove sidebar toolbar merge-down button again
As we now have the functionality available through a context menu,
we can do away with this button again, which only made sense for the
layer view and confusingly had the same icon as the move-down
button.

A button like this might make sense, but in the current architecture
it seems the toolbar buttons are the same for page- and layer-views
and thus the buttons should make sense for both, which merging
doesn't.