Ensure selected layer tracks appropriately
Setting the selected layer to the lower layer when merging and back
to the upper layer when a merge is undone.
As the TODO states, for some reason on 'redo' the selection visually
(red frame in UI) seems to return to the top layer despite the layer
ID being set appropriately.
I believe we might need `fireRebuildLayerMenu()`?