Add layer merging to 'Journal' dropdown menu
Adds an entry 'Merge Layer Down' to the 'Journal' menu, which merges
the currently selected layer with the one below.
This entry is only available when a mergable layer is selected, i.E.
any but the lowest non-background or the background layer, otherwise
it is greyed out.

This action is bound to `CTRL+M`.