Reworked PageRange to use C++17 STL.
Summary of changes:

* The function `PageRange::parse` was modified to use STL;
  - Fixes a potential bug were ranges larger than `pageCount`
    could be stored. (It is unclear if this bug could occur)
  - Eliminates a potential memory leak by avoiding `new`.

* Added documentation for `PageRange::parse`.

* `PageRangeEntry` was simplified to a small struct.
  - Its element have type `size_t` now. This matches other
    parts of the source code where pages or page counts are
    mentioned, and avoids casts between `int` and `size_t`.

* Reworked other parts of the source code to use `PageRange` with
  the new implementation.

* Added a unit test for `PageRange::parse`.

Modify PageRange::parse to reject bad input.

Exceptions are thrown if bad input is encountered.

Proper Doxygen-style documentation is added to the function.

A bug was fixed in the case of a single page input.

Changes to export dialog in page selection.

Added an "error" style when the page range is erroneous, in particular
when the PageRange::parse() function throws an invalid_argument
exception. Also, the OK button is grayed out when the page range is
invalid.

Minor improvement in GUI behavior.

The red error is hidden away when the user de-selects away from the text
field. This is appropriate behavior; otherwise even if de-selected, the
red outline remains, and may give the impression that the user should
not press OK.

Co-authored-by: Fabian Keßler <fabian_kessler@gmx.de>