Modify Tool ctor to accept reference.
We change the constructor of Tool from Tool(Tool*) to Tool(const Tool&).
This modification was a TODO that was noted in the docstring of the
constructor.

This constructor was used in ToolHandler.cpp, and so we modify the file
accordingly, passing a reference instead of a pointer where appropriate.