Improve getToolThickness.
This function previously returned an array of doubles, that the caller
would index to obtain 3 thickness widths, one for each tool type. There
was only one instance where this function was used. We replace it with a
function that takes a tool type and a tool size and returns a width
instead.

This change meant that we had to change the setSize functions in the
source code; they previously accepted an array and a tool size and had a
translation between two different kinds of ToolType enums: e.g. they
would translate STROKE_TOOL_PEN to TOOL_PEN. We instead improved the
interface of Stroke, by renaming the  previous ToolType to
StrokeToolType, to distinguish it from the ToolEnums.h ToolType, and now
Stroke has two functions, getStrokeToolType and getToolType. Finally,
the setSize functions now instead accept a mapping between ToolTypes and
thickness widths.

We also modified the docstrings a bit, and in particular removed
duplicate docstrings from EditSelectionContents.cpp and
EditSelection.cpp, and instead only retained docstrings in the
respective header files.