Add const qualifier in serialization methods.
The Serialize class has a virtual method serialize() which does not
modify the instance. The const qualifier has been added to its signature
and all the implementations of serialize(). The override keyword is also
added where appropriate.

Finally, some classes provided their own implementation of serialize()
called serializeFoo() where 'Foo' is the class name, to be able to use
the base class serialization method; we delete those serializeFoo()
methods, override the serialize() method and instead use
this->Base::serialize() where appropriate. We have the same
functionality, but the API is more C++-like.