Create drawing buffer for page once
This moves the creation of the drawing buffer of a
page to the page view. This is a preliminary fix to
get tiling up and running. In the end this should
be removed at all as it causes issues with big
resolutions. Not yet sure how I'll do that.

This fix is also buggy as hell... Most issues can
be worked around by zooming once.