Fix deadlock in SidebarPreviewPages
The deadlock is caused by `document->lock()` being called twice.
The first call is due to the lock in
`SidebarPreviewPages::updatePreviews`, which deletes the sidebar entries
and consequently waits for the "currently running job" to finish first. The
second call is due to exactly one such "currently running job" (e.g.,
RenderJob).