Fix PDF linear text selection on improper rectangles
For the purposes of this commit message, let an improper rectangle {x1,y1,x2,y2}
be one such that x1 > x2 or y1 > y2. On improper selection rectangles, the
finalized region is excessively large because 1) the selection is converted into
a proper rectangle before being passed to PopplerGlibPage::selectTextLines; and
2) the region computed from the proper rectangle is strictly larger than the
actual selection rectangle.

This commit moves the proper rectangle conversion to
PopplerGlibPage::selectTextLines, so that it correctly functions on the actual
text selection.