Load Image elements using GDK pixbuf
In summary, this commit allows images of formats other than PNG (e.g. JPEG) to
be used for Image elements.

Specifically, before this commit, the Image element stored the image data in PNG
format. Callers of the Image element API had to decode the source image and
convert to PNG before creating an Image or modifying Image data.

Now, the Image element stores (1) the raw image data and (2) the file format of
the image. This has the following benefits:

* The image can be decoded on demand (e.g., when it needs to be rendered).
* Image data (de)serialization (e.g., to clipboard) is just a copy instead of
  converting to/from PNG.
* In the future, images can be stored in their original formats instead of PNG,
  which avoids xopp file size blowup.
* Vector graphics are preserved instead of rasterized, which improves image
  quality and decreases xopp file size.

While images of arbitrary format can now be _loaded_ from an xopp file, in order
to preserve backwards compatibility, images are still converted to PNG format
when saving.