switch spec file to rpkg format
The in-tree spec file is meant for automatic snapshot builds via COPR,
which uses rpkg to parse the input spec file and create a result spec
file and srpm (containing the processed spec), builds rpms and then
distributes srpm and rpm.

The current srpm has a few problems:
- It makes COPR download the source from github (tip of master at that
  time) even though COPR cloned the repo already.
- That source is put into the srpm.
- When the rpm is built from the srpm, it queries github for the tip of
  master and uses that (and the current date) to determine the version.

As a consequence, the distributed srpm is, e.g.

xournalpp-1.1.1+dev^20220324git-1.fc35.src.rpm with
Version     : 1.1.1+dev^20220324git

and the rpm built from it is

xournalpp-1.1.1+dev^20220324gitdcad936-1.fc35.x86_64.rpm with
Version     : 1.1.1+dev^20220324gitdcad936

In particular, names and version do not coincide because the non-rpkg
spec cannot use git (see COPR's log). Also, rebuilding the rpm from the
same srpm may result in a different version if the date is different or
the master branch advanced, even though the srpm still contains the
"old" source. I.e.: That srpm does not ensure stable rebuilds and should
not be distributed.

Switching to rpkg features in the spec solves all these problems, since
the srpm is built from the cloned source (no race condition with moving
tip) and all rpkg macros are expanded when the spec is parsed, against
that one git checkout. The resulting spec and srpm are stable.

Also, we switch to `git describe` based versioning (instead of the date)
which gives a natural strictly increasing numbering, computes the base
release automatically (assuming no merge to master does strange things)
and contains the information how far off we are from the base release.

Finally, we remove `+dev` from the version string since this is not part
of any xournalpp release. The snapshot is based on (relative to) the tag
release.