Fix image loading regression (#3952)
This fixes a regression where the cairo_surface of the Image element would have swapped width and height.

As a result of the regression, the XmlImageNode uses the rendered image buffer during serialization, the incorrect cairo_surface would be saved. Existing xopp files that have been overwritten by a save have been affected by this regression.