Fix image loading failure on some images
To determine the image format, `gdk_pixbuf_loader_get_format` is invoked during
incremental image loading (to avoid having to the whole image). However, in some
cases, the format can only be determined after the image load. Because it is
assumed that the format can always be determined, this leads to assertion
failures on some users' documents.