scale small arrow heads by line length
So far, the size of arrow heads is determined purely by the line width,
which is good for medium sized and long arrows. For very small arrows,
this leads to "interesting effects", especially for double headed
arrows.

To avoid these effects, scale the arrow head by the line length for
short arrows, i.e.: make the heads no larger than 0.4 * line length.
This suites double arrows as well, and besides, 0.4 is roughly
1 - (sqrt(5) - 1) / 2 which is the golden choice.