Change float format string to use g specifier
This commit changes the float precision format string to "%.8g" to save space
when serializing.

The float precision format string is defined as "%.8f", which always prints
floats with 8 trailing digits (padding with zeros if needed). However, in many
cases, fewer than 8 trailing digits are needed. This means that by using the
"%.8g" specifier, which does not print trailing digits and may switch to
scientific notation to be shorter, the size of the serialized stroke strings
may be reduced.

As an experiment, I compared two files consisting solely of stroke data:

* File A: a 5-page, 1.4 MB document with worked out math problems.
* File B: a 16-page, 3.3 MB document full of mathematical proofs.

On File A, the total size of the stroke string data was reduced by 23.64% with
no change in the float values; this lead to a total file size reduction of
16.24% (uncompressed) and 6.40% (compressed).

On File B, the total size of the stroke string data was reduced by 23.11% with
negligible change in the float values (less than 1e-5 % relative difference);
this lead to a total file size reduction of 15.48% (uncompressed) and
18.99% (compressed).