Added functionality to the missing PDF background dialog to propose replacement file
If there is a PDF file with the same name as the missing PDF in the directory where the journal file is saved, the dialog now proposes that file to the user. If not the dialog stays as before.

This is especially useful when moving the journal file and the background to a new location. Or in my case where I switch between Linux and Windows a lot, the background can be automatically found even tho Linux and Windows are using a different path structure.