feat: Min size for shape recognizer now customizable
For some users, using the shape recognizer was problematic because it
activated when writing letters. Now, users can set a minimum "width" and
"height". For example, in the case of "width", if the minimum x value of
the stroke is subtracted from the maximum x value of the stroke, and if
that is less than the configured "width", then the shape will not be
drawn.