restructure: Control minimum distance with a single property
The X-Y-based controls didn't make much sense. It was harder to use for
the end user, and it would not recognize shapes in some situations when
it should. For example, if the x-axis is super long (larger than min
distance), and the y-axis is short (shorter than the min distance), it
wouldn't use the recognized shape because only one of the axis is shorter
than the min-distance. With a single control, this has been fixed. Now,
both X and Y directions need to be smaller than the minimum distance to
not use the recognized shape.