Fix PDF toolbox stroke actions when PDF page num is not current page
Previously, `PdfFloatingToolbox::createStrokes` would incorrectly no-op if the
page number (of the document) is not equal to the page number of the background
PDF that the text selection is over.

This commit changes the fix to instead retrieve the PDF page number of the
current document page, and compare that against the PDF page number of the text
selection.