Add LoadHandler::fixNullPressureValues
This adds a filter in LoadHandler to ensure that the pressure values of
every point in every stroke is positive (or Point::NO_PRESSURE if the
stroke is not pressure-sensitive).
All points with a non-positive pressure value are removed, and the corresponding stroke is cut.
An invalid stroke can thus become several valid ones (or none at all).

A suitable test unit is also added