Fix PDF toolbox showing at the wrong position on Wayland
Originally, the PDF toolbox position was computed with respect to the main
application window's origin. However, we use the get-child-position signal of
GtkOverlay (PdfFloatingToolbox::getOverlayPosition) to compute the position,
which must be relative to the GtkOverlay widget's origin. This means that if
overlay has a non-zero position with respect its parent window, then the PDF
toolbox will be displayed at an incorrect offset. Such a non-zero position can
occur when using Wayland.

This commit fixes the calculation by transforming the window coordinates to the
overlay coordinates in PdfFloatingToolbox::show().