Do not define in `XML_VERSION_STR` variable InputStreamException.cpp file
We do not use this value in the InputStreamException.cpp and provide it in the
Serializable.h using the `extern` declaration.
Considering that we can be sure the serializing utils are always up to date
when rebuilding the project, there is no problem with having this value header-only
for the few usages.

While at it, the value can be made const (as clang-tidy suggested).