Fix error handling in "Annotate PDF" menu option
Previously, the "Annotate PDF" menu option works by closing the current document
and then creating a new document from the given PDF. However, if the given "PDF"
fails to load correctly, then the application will be left with no loaded
document. This violates the invariant that a document must always be loaded in
the application, causes crashes to occur on e.g. tool changes.

This commit fixes the bug by first loading the PDF into a dummy document. If the
load is successful, then the current document instance is overwritten with the
dummy document. The approach is similar to the one in Control::openFile().