Add plugin to fit the page size to the content/selection
The idea behind this, is to be able to prototype small images for e.g.
LaTeX documents (quick and dirty drawing before probably time consuming
tikzing the drawing).

Caveat: Only works with strokes, no images and no text boxes are taken
into account. Might change in the future, but for this we need
get/add api-functions for these types.