[luaAPI] refactor error handling
Standardize error handling for the luApi. Real errors are thrown as
`luaL_error` so that it is a real error in the lua script (which can be
catched with `pcall`).
In cases where the behaviour was to be expected (e.g. ressource is
unavailable), `nil, errMsg` will be returned (currently no example
besides `applib_glib_rename`). `g_warning`s are avoided for noticing
errors as they cannot be handled in the lua script.
This convention is documented in the head of the `luapi_application.h`.

Also as `return luaL_error();` is more ideomatic than (`luaL_error();
return;`), the first one is preferred (no difference in behaviour).

The error message when loading a lua script failed was refactored as
well, now there is a distinction if *loading* or *running* failed. And
when *loading* failed, the error message is shown as well.