scale arrow heads differently in more regimes
Depending on the ratio between linelength and thickness, different
choices of scalings show different artifacts. To avoid the worst ones,
choose more regimes for scaling (partly suggested by @tattsan).
For very "non-slim" arrows, we choose the cut-off depending on the
number of arrow heads so that the directionality remains visible when
shrinking down to a point. The resulting shape (arrowdist) is still a
linear function of (thickness, lineLength).

Suggested-By: SATO Tatsuya <tattsan@users.noreply.github.com>